refactor(settings): extract AccountSwitcher widget
Move the multi-account switcher from settings_screen.dart into a
standalone AccountSwitcher widget under lib/features/settings/widgets/.
The widget owns its own visibility gate (hasMultipleAccounts) so the
call site is a single line. No visual or behavioral change.

Closes #55
